requirements.txt
flask 
request 
python-dotenv 
duckduckgo-search
phidata
groq